Extract focus assertions
